requirements.txt
linode-cli==5.36.0
PyYAML==6.0
requests==2.28.2
